+++ pytorch_lightning/plugins/__init__.py
@@ -22,7 +22,7 @@ from pytorch_lightning.plugins.precision.tpu_bf16 import TPUBf16PrecisionPlugin
 from pytorch_lightning.plugins.training_type.ddp import DDPStrategy
 from pytorch_lightning.plugins.training_type.ddp2 import DDP2Plugin
 from pytorch_lightning.plugins.training_type.ddp_spawn import DDPSpawnPlugin
-from pytorch_lightning.plugins.training_type.deepspeed import DeepSpeedPlugin
+from pytorch_lightning.plugins.training_type.deepspeed import DeepSpeedStrategy
 from pytorch_lightning.plugins.training_type.dp import DataParallelPlugin
 from pytorch_lightning.plugins.training_type.fully_sharded import DDPFullyShardedStrategy
 from pytorch_lightning.plugins.training_type.horovod import HorovodStrategy
@@ -48,7 +48,7 @@ __all__ = [
     "DDPStrategy",
     "DDPSpawnPlugin",
     "DDPFullyShardedStrategy",
-    "DeepSpeedPlugin",
+    "DeepSpeedStrategy",
     "DeepSpeedPrecisionPlugin",
     "DoublePrecisionPlugin",
     "HorovodStrategy",